Refactor to create the BigQueryService inside AppComponents.scala, and use the existing configuration object
