+++ app/controllers/BanditDataController.scala
@@ -3,8 +3,9 @@ package controllers
 import com.gu.googleauth.AuthAction
 import com.typesafe.scalalogging.LazyLogging
 import io.circe.syntax.EncoderOps
+import models.Channel
 import play.api.mvc.{AbstractController, Action, ActionBuilder, AnyContent, ControllerComponents, Result}
-import services.DynamoBanditData
+import services.{BigQueryService, DynamoBanditData}
 import utils.Circe.noNulls
 import zio.{IO, ZEnv, ZIO}
 
@@ -17,6 +18,7 @@ class BanditDataController(
   stage: String,
   runtime: zio.Runtime[ZEnv],
   dynamo: DynamoBanditData,
+  bigQueryService: BigQueryService,
 )(implicit ec: ExecutionContext) extends AbstractController(components) with LazyLogging {
 
   private def run(f: => ZIO[ZEnv, Throwable, Result]): Future[Result] =
@@ -34,5 +36,27 @@ class BanditDataController(
         .getDataForTest(testName, channel, sampleCount)
         .map(data => Ok(noNulls(data.asJson)))
     }
+  }
+
+  def getLTVData(testName: String, channel: Channel, bigQueryService: BigQueryService) = {
+    logger.info(s"Start BigQuery testing")
+
+        val projectId = s"datatech-platform-${stage.toLowerCase}"
+        val query = bigQueryService.buildQuery(testName, channel, stage.toLowerCase);
+
+        val result = bigQueryService.runQuery(query, projectId) match {
+          case Left(error) =>
+            Left(error)
+          case Right(results) =>
+            val bigQueryResult = bigQueryService.getBigQueryResult(results)
+            logger.info(s"BigQueryResult: $bigQueryResult");
+            Right(bigQueryResult)
+
+        }
+        logger.info(s"Result: $result");
+
+    logger.info(s"End BigQuery testing")
+
+
   }
 }
